:shit: Rename function to fix deployment issue
Renaming the function in the hope it fixes the issue with the
orchestration throwing an error when it can't find the named function.
Even though locally the function exists and is correctly named, when it
is deployed to Azure the function has a capital `O` i.e. `Orchestration`
as opposed to `orchestration`. Hopefully the full renaming of the
function will sort out the deployment issue.