:closed_book: Add clarity as to what closing the modal does
There was a tendency for people to close the modal by clicking on the
'close' button almost as a response to seeing the button to click and
clicking it rather than out of a real wish to close the modal. This
wasn't actually what user's wanted to happen. In order to reduce the
likelihood of this happening the button has been updated to say 'cancel'
which is more accurate of what happens when the button is clicked as the
request is not possible to proceed without the modal being open.